Fix squeezing word clusters without counterparts
If learners add a lot of wrong words, these can lack a counterpart
in the solution, and clusters between matches can be squeezed together
in order to not count these counterpartless words as mistakes.

The previous implementation did this squeezing for the whole
sentence and could lead to wrong word alignment.